rename bytebuf mod
