+++ pumpkin-protocol/src/server/login/plugin_response.rs
@@ -1,6 +1,6 @@
 use crate::{
     ServerPacket, VarInt,
-    bytebuf::{ByteBuf, ReadingError},
+    ser::{ByteBuf, ReadingError},
 };
 use bytes::{Buf, Bytes};
 use pumpkin_data::packet::serverbound::LOGIN_CUSTOM_QUERY_ANSWER;
