+++ pumpkin-protocol/src/server/play/cookie_response.rs
@@ -1,7 +1,7 @@
 use crate::{
     ServerPacket, VarInt,
-    bytebuf::{ByteBuf, ReadingError},
     codec::identifier::Identifier,
+    ser::{ByteBuf, ReadingError},
 };
 use bytes::Buf;
 use pumpkin_data::packet::serverbound::PLAY_COOKIE_RESPONSE;
