+++ pumpkin-protocol/src/server/play/player_command.rs
@@ -4,7 +4,7 @@ use pumpkin_macros::packet;
 
 use crate::{
     ServerPacket, VarInt,
-    bytebuf::{ByteBuf, ReadingError},
+    ser::{ByteBuf, ReadingError},
 };
 
 #[packet(PLAY_PLAYER_COMMAND)]
