+++ pumpkin/src/error.rs
@@ -1,6 +1,6 @@
 use log::log;
 use pumpkin_inventory::InventoryError;
-use pumpkin_protocol::bytebuf::ReadingError;
+use pumpkin_protocol::ser::ReadingError;
 use std::fmt::Display;
 
 pub trait PumpkinError: Send + std::error::Error + Display {
